Update README.md
This research proposed a wheelchair-person fall detection
system with IOT which is cost-effective and reliable to detect
falls and alert surrounding to call for help. For fall detection,
a gyroscope, GPS module, FSR pressure sensor and microcontroller are implemented into the system. The gyroscope is used to detect the position of the wheelchair while the FSR
pressure sensor which is placed on the seat pad of the wheelchair
will be used to detect and recognise the gesture of the user.
Both work together to detect fall events which increases the
accuracy of the fall detection. GPS module used to allocate
the location of the wheelchair when a fall event occurs. When
a fall event occurs, all the data including the location of the
wheelchair will be sent to the Blynk mobile application. The IOT
system will send email notifications to the registered person
to alert them of fall events and the help needed. Moreover,
this system requires less implementation cost and provides
a quick response. It can be installed in the existing commercial
wheelchair. However, the limitation of this system is it requires
a good WIFI connection. For future recommendations, a better
GUI design can be implemented into the system besides
that, more detection systems can be added to increase the
accuracy of the system.